requirement.txt
astroid==2.1.0
docopt==0.6.2
et-xmlfile==1.0.1
ExifRead==2.1.2
fortnum==0.0.9
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
openpyxl==2.5.12
openpyxl-templates==0.1.24
packbits==0.6
Pillow==5.4.0
psd-tools2==1.7.27
pylint==2.2.2
six==1.12.0
typed-ast==1.1.1
wrapt==1.10.11
